backend/requirements.txt
# ValuHub 后端依赖
# FastAPI + PostgreSQL + Redis + 其他依赖

fastapi==0.104.1
uvicorn[standard]==0.23.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
redis==5.0.1
celery==5.3.4
weasyprint==60.1
oss2==2.18.4
